+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - Ensure Apple Silicon-specific installers are not listed
   - ie. M2 specific build (21F2092)
 - Avoid adding OpenCore icon in boot picker if Windows bootloader on same partition
+- Add error-handling to corrupt/non-standard NVRAM variables
+- Add warning prompt when using 'Allow native models'
+  - Attempt to avoid misuse of option
 - Ventura Specific Updates:
   - Switch boot.efi model patch to iMac18,1
   - Resolve pre-Force Touch Trackpad support in Ventura